feat(theme): ringed sidebar tabs in the two cyberpunk presets
Both take the same shape -- sidebar_tab_flat_edge = "none", so every corner
follows tab_radius and the outline closes the whole way round. They differ in
one thing, which is what makes them a pair worth having:

  cyberpunk_neon   active tab ringed in its focus cyan, the rest bare
                   (sidebar_tab_border_color transparent)
  cyberpunk_edge   every tab ringed -- muted violet inactive, amber active,
                   the same treatment it already gives its card outline and
                   the bottom rule it exists to demonstrate

Both pin sidebar_indicator_width to the ring's width, and that is load-bearing
rather than tidiness. The strip shares the ring's content-facing edge and is
painted under it, so at equal widths the ring covers it exactly and the tab is
one clean line; left at the 3px default it stuck out inside the ring --
measured a pink sliver at x=23 against neon's cyan, and 3px of amber where
edge draws 1.5. A test samples one pixel further in than the ring can reach
and requires the tab's own fill there, so either preset drifting reopens it.

SIDEBAR.indicator_width is now a float, like the dock tabs' own
indicator_width, which is what lets cyberpunk_edge run 1.5 throughout as it
does everywhere else. ThemeSpec.sidebar_indicator_width and the JSON field
follow.

_inked_edges (test and smoke) now zeroes the strip for both of its renders.
It measures the outline by turning it off, and where the two share a colour --
edge rings and stripes its active tab in the same amber -- the strip went on
painting that edge with the outline gone, reading as an edge the outline never
drew.

408 tests pass; the eight new preset assertions were sabotage-verified, as was
the guard that no third preset picks the shape up by accident.
smoke_sidebar_paint.py checks both presets on real Win32; run_all.py is green.

+++ lace/dock_theme.py
@@ -282,7 +282,9 @@ class DockSidebarStyleSchema:
 
     # Highlights & Badges
     indicator_color: Optional[List[int]] = None
-    indicator_width: int = 3
+    # Float like the dock tabs' own indicator_width, so a theme can match the
+    # strip to a fractional outline width — a mismatch steps the edge they share.
+    indicator_width: float = 3.0
     indicator_position: str = "right"  # "left" or "right"
 
     badge_bg: Optional[List[int]] = None
@@ -436,7 +438,9 @@ class ThemeSpec:
     #: Width and edge of the sidebar tab's highlight strip — the counterpart of
     #: ``indicator_width`` / ``indicator_position`` for the dock widget tabs.
     #: ``"left"`` is the window-facing edge, ``"right"`` the content-facing one.
-    sidebar_indicator_width: Optional[int] = None
+    #: Give it the same width as the outline: the strip sits *on* one of the
+    #: outline's edges, so a mismatch steps that edge thicker.
+    sidebar_indicator_width: Optional[float] = None
     sidebar_indicator_position: Optional[str] = None
 
     # Tooltip colors — when omitted, derived from the panel/text seed colors.
@@ -542,7 +546,7 @@ def _build_theme(
     sidebar_tab_border_color: Optional[list] = None,
     sidebar_tab_border_active_color: Optional[list] = None,
     sidebar_tab_border_closed: Optional[bool] = None,
-    sidebar_indicator_width: Optional[int] = None,
+    sidebar_indicator_width: Optional[float] = None,
     sidebar_indicator_position: Optional[str] = None,
     tooltip_bg: Optional[list] = None,
     tooltip_text: Optional[list] = None,
